api/requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
celery==5.2
Django==3.2.8
django-filter==21.1
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-storages==1.12.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-spectacular==0.20.2
firebase-admin==3.2.1
h11==0.12.0
idna==3.3
inflection==0.5.1
jsonschema==4.2.1
phonenumbers==8.12.35
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.18.0
python-magic==0.4.24
pytz==2021.3
PyYAML==6.0
pymssql==2.2.2
requests==2.26.0
sqlparse==0.4.2
qrcode==7.3.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
